Pulling dev into assets-module, i needed the entity's to line up so i can make the auditLog

+++ src/main/java/com/example/demo/controller/AssetController.java
@@ -39,7 +39,7 @@ public class AssetController {
 
     // search by ID //4//DONE
     @GetMapping("/{id}")
-    public Asset getAssetById(@PathVariable Integer id){
+    public Asset getAssetById(@PathVariable Long id){
         return service.getAssetById(id);
     }
 
@@ -100,70 +100,70 @@ public class AssetController {
     // update entire asset //14//DONE
     @PutMapping("/{id}")
     public Asset updateAsset(
-            @PathVariable Integer id,
+            @PathVariable Long id,
             @RequestBody Asset asset) {
         return service.editAsset(id, asset);
     }
 
     // update title //15//DONE
     @PatchMapping("/{id}/title")
-    public Asset updateTitle(@PathVariable Integer id, @RequestParam String title) {
+    public Asset updateTitle(@PathVariable Long id, @RequestParam String title) {
         return service.updateTitle(id, title);
     }
 
     // update category //16//DONE
     @PatchMapping("/{id}/category")
-    public Asset updateCategory(@PathVariable Integer id, @RequestParam String category) {
+    public Asset updateCategory(@PathVariable Long id, @RequestParam String category) {
         return service.updateCategory(id, category);
     }
 
     // update serial //17//DONE
     @PatchMapping("/{id}/serial")
-    public Asset updateSerial(@PathVariable Integer id, @RequestParam String serialNumber) {
+    public Asset updateSerial(@PathVariable Long id, @RequestParam String serialNumber) {
         return service.updateSerialNumber(id, serialNumber);
     }
 
     // update acquisition-date //18//DONE
     @PatchMapping("/{id}/acquisition-date")
-    public Asset updateAcquisitionDate(@PathVariable Integer id, @RequestParam LocalDate date) {
+    public Asset updateAcquisitionDate(@PathVariable Long id, @RequestParam LocalDate date) {
         return service.updateAcquisitionDate(id, date);
     }
 
     // update cost //19//DONE
     @PatchMapping("/{id}/cost")
-    public Asset updateCost(@PathVariable Integer id, @RequestParam BigDecimal cost) {
+    public Asset updateCost(@PathVariable Long id, @RequestParam BigDecimal cost) {
         return service.updateCost(id, cost);
     }
 
     // update location //20//DONE
     @PatchMapping("/{id}/location")
-    public Asset updateLocation(@PathVariable Integer id, @RequestParam String location) {
+    public Asset updateLocation(@PathVariable Long id, @RequestParam String location) {
         return service.updateLocation(id, location);
     }
 
     // update condition//21//DONE
     @PatchMapping("/{id}/condition")
-    public Asset updateCondition(@PathVariable Integer id, @RequestParam String condition) {
+    public Asset updateCondition(@PathVariable Long id, @RequestParam String condition) {
         return service.updateCondition(id, condition);
     }
 
     //update photo //22//DONE
     @PatchMapping("/{id}/photo")
-    public Asset updatePhoto(@PathVariable Integer id, @RequestParam String photoPath) {
+    public Asset updatePhoto(@PathVariable Long id, @RequestParam String photoPath) {
         return service.updatePhotoPath(id, photoPath);
     }
 
     // update status of asset//23//DONE
     @PatchMapping("/{id}/status")
     public Asset updateStatus(
-            @PathVariable Integer id,
+            @PathVariable Long id,
             @RequestParam Asset.Status status) {
         return service.updateStatus(id, status);
     }
 
     // delete asset //24//DONE but when deleting for example id 1, id 2 does not become id 1 and so forth
     @DeleteMapping("/{id}")
-    public void deleteAsset(@PathVariable Integer id) {
+    public void deleteAsset(@PathVariable Long id) {
         service.deleteAsset(id);
     }
